services/llm-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx>=0.25.0
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
anthropic>=0.40.0
google-generativeai>=0.8.0
